# edge_cloud_management_api/specification/openapi.yaml
...
openapi: 3.0.3
info:
  title: Edge Application Management API
  version: 0.9.3-wip
  description: |
    Edge Application Management API allows API consumers to manage the
    Life Cycle of an Application and to Discover Edge Cloud Zones.
    # Overview
    The reference scenario foresees a distributed Telco Edge Cloud where any
    Application Delevoper, known as an Application Provider, can host and
    deploy their application according to their specifications and operational
    criteria (e.g. within an specific geographical zone for data protection
    purposes, ensure a minimum QoS for the application clients, etc).
    Through Telco Edge Cloud services Developers around the globe can be
    benefit from the traditional Cloud strengths but expertise and advantages
    of the Mobile Network Operators offering to their users an evolved
    experience for XR, V2X, Holographic and other new services.

    # Introduction
    The Edge Application Management API provides capabilities for lifecycle
    management of application, instances and edge cloud zone discovery.
    Lifecycle Management allows Application Provider to onboard
    their application to the Edge Cloud Platform which do bookkeeping,
    resource validation and other pre-deployment operations.
    Application details can contain components network specification,
    package type (QCOW2, OVA, CONTAINER, HELM), operating system details and
    respository to download the image of the desired application.
...

...

...

...

...

...

...

...

...
paths:
  /apps:
    post:
      # security:
      #   - openId:
      #       - edge-application-management:apps:write
      tags:
        - Application
      summary: Submit application metadata to the Edge Cloud Provider.
      description: |
        Contains the information about the application to be
        instantiated in the Edge Cloud
      operationId: edge_cloud_management_api.controllers.app_controllers.submit_app
      parameters:
        - $ref: "#/components/parameters/x-correlator"
      requestBody:
        description: |
          The Application Provider request contains mandatory
          criteria (e.g. required CPU, memory, storage, bandwidth) and
          optional parameters.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppManifest"
            examples:
              KubernetesExample:
                summary: Example for Kubernetes app
                value:
                  name: nginx_web_app
                  appProvider: nginx_inc
                  version: 1.0.0
                  packageType: HELM
                  operatingSystem:
                    architecture: x86_64
                    family: UBUNTU
                    version: OS_VERSION_UBUNTU_2204_LTS
                    license: OS_LICENSE_TYPE_FREE
                  appRepo:
                    type: PRIVATEREPO
                    imagePath: "https://charts.bitnami.com/bitnami/nginx:1.25.2"
                    userName: helm-user
                    credentials: secure-token
                    authType: DOCKER
                    checksum: sha256:4d67e4c1a4b4c8...
                  requiredResources:
                    infraKind: kubernetes
                    applicationResources:
                      cpuPool:
                        numCPU: 2
                        memory: 4096
                        topology:
                          minNumberOfNodes: 3
                          minNodeCpu: 2
                          minNodeMemory: 4096
                      gpuPool:
                        numCPU: 4
                        memory: 8192
                        gpuMemory: 32
                        topology:
                          minNumberOfNodes: 2
                          minNodeCpu: 4
                          minNodeMemory: 8192
                          minNodeGpuMemory: 16
                    isStandalone: false
                    version: "1.29"
                    additionalStorage: 100GB
                    networking:
                      primaryNetwork:
                        provider: cilium
                        version: "1.13"
                      additionalNetworks:
                        - name: net1
                          interfaceType: vfio-pci
                        - name: backend-net
                          interfaceType: netdevice
                        - name: db-net
                          interfaceType: interface
                    addons:
                      monitoring: true
                      ingress: true
                  componentSpec:
                    - componentName: nginx_server
                      networkInterfaces:
                        - interfaceId: eth0
                          protocol: TCP
                          port: 80
                          visibilityType: VISIBILITY_EXTERNAL
                        - interfaceId: eth1
                          protocol: TCP
                          port: 443
                          visibilityType: VISIBILITY_EXTERNAL

        required: true
      responses:
        "201":
          description: Application created successfully
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmittedApp"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
              example:
                status: 409
                code: CONFLICT
                message: "App already exists"
        "500":
          $ref: "#/components/responses/500"
        "501":
          $ref: "#/components/responses/501"
        "503":
          $ref: "#/components/responses/503"
    get:
      # security:
      #   - openId:
      #       - edge-application-management:apps:read
      tags:
        - Application
      summary: Retrieve a list of existing Applications
      description: |
        Get the list of all existing Application definitions from the
        Edge Cloud Provider that the user has permission to view.
      operationId: edge_cloud_management_api.controllers.app_controllers.get_apps
      parameters:
        - $ref: "#/components/parameters/x-correlator"
      responses:
        "200":
          description: List of existing applications
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppManifest"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
  /apps/{appId}:
    get:
      # security:
      #   - openId:
      #       - edge-application-management:apps:read
      tags:
        - Application
      summary: Retrieve the information of an Application
      description: |
        Ask the Edge Cloud Provider the information for a given application
      operationId: edge_cloud_management_api.controllers.app_controllers.get_app
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appId
          description: |
            A globally unique identifier associated with the
            application.
            Edge Cloud Provider generates this identifier when the application
            is submitted.
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AppId"
      responses:
        "200":
          description: Information of Application
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                type: object
                properties:
                  appManifest:
                    $ref: "#/components/schemas/AppManifest"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
    delete:
      # security:
      #   - openId:
      #       - edge-application-management:apps:delete
      tags:
        - Application
      summary: |
        Delete an Application from an Edge Cloud Provider
      description: Delete all the information and content
        related to an Application
      operationId: edge_cloud_management_api.controllers.app_controllers.delete_app
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appId
          in: path
          description: |
            Identificator of the application to be
            deleted provided by the Edge Cloud Provider
            once the submission was successful
          required: true
          schema:
            $ref: "#/components/schemas/AppId"
      responses:
        "202":
          description: Request accepted
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
        "204":
          description: App deleted
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
              example:
                status: 409
                code: CONFLICT
                message: "App with a running application instance
                  cannot be deleted"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
  /appinstances:
    post:
      # security:
      #   - openId:
      #       - edge-application-management:instances:write
      tags:
        - Application
      summary: Instantiation of an Application
      description: |
        Ask the Edge Cloud Platform to instantiate an application to one
        or several Edge Cloud Zones with an Application as an input and an
        Application Instance as the output.
      operationId: edge_cloud_management_api.controllers.app_controllers.create_app_instance
      parameters:
        - $ref: "#/components/parameters/x-correlator"
      requestBody:
        description: |
          Information about the application and where to deploy it.
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - appId
                - edgeCloudZoneId
              properties:
                name: 
                  $ref: '#/components/schemas/AppInstanceName'
                appId:
                  $ref: "#/components/schemas/AppId"
                edgeCloudZoneId:
                  $ref: "#/components/schemas/EdgeCloudZoneId"
                kubernetesClusterRef:
                  $ref: "#/components/schemas/KubernetesClusterRef"
        required: true
      responses:
        "202":
          description: Application instantiation accepted
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
            Location:
              description: Contains the URI of the newly created application.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  appInstances:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppInstanceInfo"
                minItems: 1
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
              example:
                status: 409
                code: CONFLICT
                message: "Application already instantiated in the given
                  Edge Cloud Zone or Edge Cloud Region"
        "500":
          $ref: "#/components/responses/500"
        "501":
          $ref: "#/components/responses/501"
        "503":
          $ref: "#/components/responses/503"
    get:
      # security:
      #   - openId:
      #       - edge-application-management:instances:read
      tags:
        - Application
      summary: Retrieve the information of Application Instances for a given App
      description: |
        Ask the Edge Cloud Provider the information of the instances for a
        given application
      operationId: edge_cloud_management_api.controllers.app_controllers.get_app_instance
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appId
          description: |
            A globally unique identifier associated with
            the application.
            Edge Cloud Provider generates this identifier when the
            application is submitted.
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AppId"
        - name: appInstanceId
          description: |
            A globally unique identifier associated with a running
            instance of an application within an specific Edge Cloud Zone.
            Edge Cloud Provider generates this identifier.
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AppInstanceId"
        - name: region
          description: |
            Human readable name of the geographical Edge Cloud Region of
            the Edge Cloud. Defined by the Edge Cloud Provider.
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EdgeCloudRegion"
      responses:
        "200":
          description: Information of Application Instances
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                type: object
                properties:
                  appInstanceInfo:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppInstanceInfo"
                minItems: 1
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
  /appinstances/{appInstanceId}:
    delete:
      # security:
      #   - openId:
      #       - edge-application-management:instances:delete
      tags:
        - Application
      summary: Terminate an Application Instance
      description: |
        Terminate a running instance of an application within
        an Edge Cloud Zone
      operationId: edge_cloud_management_api.controllers.app_controllers.delete_app_instance
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appInstanceId
          in: path
          description: |
            Identificator of the specific application instance
            that will be terminated
          required: true
          schema:
            $ref: "#/components/schemas/AppInstanceId"
      responses:
        "202":
          description: |
            Request accepted to be processed. It applies for async
            deletion process
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
        "204":
          description: Application Instance Deleted
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
...
components:
...
  parameters:
    x-correlator:
      name: x-correlator
      in: header
      description: |
        Correlation id for the different services
      schema:
        type: string
  headers:
    x-correlator:
      description: |
        Correlation id for the different services
      required: false
      schema:
        type: string
        format: uuid
  schemas:
...
    AppId:
      type: string
      format: uuid
      description: |
        A globally unique identifier associated with the application.
        Edge Cloud Platform generates this identifier when the
        Application is submitted.
...
    AppInstanceName:
      type: string
      pattern: ^[A-Za-z][A-Za-z0-9_]{1,63}$
      description: Name of the App instance, scoped to the AppProvider
    AppInstanceInfo:
      description: Information about the application instance.
      type: object
      properties:
        appInstanceId:
          $ref: "#/components/schemas/AppInstanceId"
        status:
          description: Status of the application instance (default is 'unknown')
          type: string
          enum:
            - ready
            - instantiating
            - failed
            - terminating
            - unknown
          default: unknown
        componentEndpointInfo:
          description: |
            Information about the IP and Port exposed by the
            Edge Cloud Platform.
            Application Client shall use these access points to reach this
            application instance
          type: array
          items:
            type: object
            required:
              - interfaceId
              - accessPoints
            properties:
              interfaceId:
                type: string
                pattern: ^[A-Za-z0-9][A-Za-z0-9_]{6,30}[A-Za-z0-9]$
                description: |
                  This is the interface Identifier that Application Provider
                  defines when application is being submitted.
              accessPoints:
                $ref: "#/components/schemas/AccessEndpoint"
          minItems: 1
        kubernetesClusterRef:
          $ref: "#/components/schemas/KubernetesClusterRef"
        edgeCloudZone:
          $ref: "#/components/schemas/EdgeCloudZone"
...
    EdgeCloudZoneId:
      type: string
      format: uuid
      description: |
        Unique identifier created by the Edge Cloud Platform to identify an
        Edge Cloud Zone within an Edge Cloud.
...
    ErrorInfo:
      type: object
      description: Information about the error
      properties:
        status:
          type: integer
          description: HTTP status code returned along with this error response
        code:
          type: string
          description: Code given to this error
        message:
          type: string
          description: Detailed error description
      required:
        - status
        - code
        - message
...
    KubernetesClusterRef:
      description: |
        A global unique identifier associated with a Kubernetes cluster
        infrastructure.
      type: string
      format: uuid
      example: "642f6105-7015-4af1-a4d1-e1ecb8437abc"
...
    SubmittedApp:
      description: Information about the submitted app
      type: object
      properties:
        appId:
          $ref: "#/components/schemas/AppId"
...
  responses:
    "400":
      description: Bad request
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 400
            code: INVALID_ARGUMENT
            message: "Schema validation failed at  ..."
    "401":
      description: Unauthorized
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 401
            code: UNAUTHENTICATED
            message: "Authorization failed: ..."
    "403":
      description: Unauthorized
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 403
            code: PERMISSION_DENIED
            message: "Operation not allowed: ..."
    "404":
      description: Not Found
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 404
            code: NOT_FOUND
            message: "Resource does not exist"
    "500":
      description: Internal Server Error
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 500
            code: INTERNAL
            message: "Internal server error: ..."
    "501":
      description: Not Implemented
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 501
            code: NOT_IMPLEMENTED
            message: "Service not implemented"
    "503":
      description: Service Unavailable
      headers:
        x-correlator:
          $ref: "#/components/headers/x-correlator"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
          example:
            status: 503
            code: UNAVAILABLE
            message: "Service unavailable"